backend project restructuring
