--- server/agentauthapp/urls.py
@@ -1,8 +1,0 @@
-# dbapp/urls.py
-from django.urls import path
-from . import views
-
-
-urlpatterns = [
-    path('token/', views.generate_access_token, name="generate_access_token"),
-]
