+++ server/core/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class AgentauthappConfig(AppConfig):
+class DbappConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'agentauthapp'
+    name = 'core'
