+++ server/core/views/user_views.py
@@ -1,0 +1,8 @@
+# user_views.py
+
+from django.http import JsonResponse
+from ..models.user import User
+from django.shortcuts import get_object_or_404
+import json
+
+# Add user-related views here
