+++ server/core/views/userdetails_views.py
@@ -1,0 +1,8 @@
+# userdetails_views.py
+
+from django.http import JsonResponse
+from ..models.user_details import UserDetails
+from django.shortcuts import get_object_or_404
+import json
+
+# Add user details related views here
