+++ server/server/settings.py
@@ -47,9 +47,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'corsheaders',
-    'dbapp',
-    'agentauthapp',
-
+    'core',
 ]
 
 MIDDLEWARE = [
